Separate repo resume functions into reusable functions in the sources package
